requirements.txt
setuptools
pillow
pycairo
cairosvg
numpy
perlin_noise
pytest